requirements.txt
streamlit
yfinance
pandas_ta
google-generativeai
requests